+++ src/frame/getFrameMessage.test.ts
@@ -1,4 +1,3 @@
-import { mockNeynarResponse } from './mock';
 import { getFrameMessage } from './getFrameMessage';
 import { getMockFrameRequest } from './getMockFrameRequest';
 import { neynarBulkUserLookup } from '../utils/neynar/user/neynarUserFunctions';
@@ -16,6 +15,30 @@ jest.mock('../utils/neynar/frame/neynarFrameFunctions', () => {
     neynarFrameValidation: jest.fn(),
   };
 });
+
+function mockNeynarResponse(
+  fid: number,
+  addresses: string[] | undefined,
+  lookupMock: jest.Mock,
+  frameValidationMock: jest.Mock = jest.fn(),
+) {
+  const neynarResponse = {
+    users: [
+      {
+        verifications: addresses,
+      },
+    ],
+  };
+  lookupMock.mockResolvedValue(neynarResponse);
+
+  frameValidationMock.mockResolvedValue({
+    valid: true,
+    interactor: {
+      fid,
+      verified_accounts: addresses,
+    },
+  });
+}
 
 describe('getFrameValidatedMessage', () => {
   it('should return undefined if the message is invalid', async () => {
